feat: renaming packages and creating test for new tail-recursive function
