+++ src/main/kotlin/math/Factorial.kt
@@ -1,4 +1,4 @@
-package mathematics
+package math
 
 import java.security.InvalidParameterException
